lab 6 done fixed
