--- src/main/java/com/mycompany/app/Player.java
@@ -1,8 +1,0 @@
-package com.mycompany.app;
-
-public class Player {
-    public char symbol;
-    public int move;
-    public boolean selected;
-    public boolean win;
-}
