--- src/main/java/com/mycompany/app/Point.java
@@ -1,9 +1,0 @@
-package com.mycompany.app;
-
-public class Point {
-    public int row, col;
-    public Point(int row, int col) {
-        this.row = row;
-        this.col = col;
-    }
-}
