--- src/main/java/com/mycompany/app/State.java
@@ -1,5 +1,0 @@
-package com.mycompany.app;
-
-public enum State { 
-    PLAYING, OWIN, XWIN, DRAW 
-}
